share code between rhsso and keycloak deployments
