+++ pkg/model/keycloak_deployment.go
@@ -1,8 +1,6 @@
 package model
 
 import (
-	"strings"
-
 	"github.com/keycloak/keycloak-operator/pkg/apis/keycloak/v1alpha1"
 	v13 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
@@ -316,31 +314,6 @@ func KeycloakDeploymentReconciled(cr *v1alpha1.Keycloak, currentState *v13.State
 	}
 	reconciled.Spec.Template.Spec.InitContainers = KeycloakExtensionsInitContainers(cr)
 	return reconciled
-}
-
-func KeycloakExtensionsInitContainers(cr *v1alpha1.Keycloak) []v1.Container {
-	return []v1.Container{
-		{
-			Name:  "extensions-init",
-			Image: KeycloakInitContainerImage,
-			Env: []v1.EnvVar{
-				{
-					Name:  KeycloakExtensionEnvVar,
-					Value: strings.Join(cr.Spec.Extensions, ","),
-				},
-			},
-			VolumeMounts: []v1.VolumeMount{
-				{
-					Name:      "keycloak-extensions",
-					ReadOnly:  false,
-					MountPath: KeycloakExtensionsInitContainerPath,
-				},
-			},
-			TerminationMessagePath:   "/dev/termination-log",
-			TerminationMessagePolicy: "File",
-			ImagePullPolicy:          "Always",
-		},
-	}
 }
 
 func KeycloakVolumeMounts(extensionsPath string) []v1.VolumeMount {
